Crate rfsapi

Raw Filesystem API for Rust — enable simpler browsing with ease

This library is to enable both servers and clients to use the RFSAPI, see D'Oh for usage example.

Modules

util
Module containing various utility functions.

Structs

FilesetData
Basic RFSAPI response returned by the server.

RawFileData
Information about a file available through RFSAPI.

RawFsApiHeader
Header to specify when doing a request for the Raw Filesystem API, designated by "X-Raw-Filesystem-API".